+++ lib/module/I2CModule.h
@@ -9,5 +9,5 @@ struct I2CConfig {
 
 class I2CModule : public Module {
 public:
-    I2CModule(std::string name, unsigned long cycleCheckTime = DEFAULT_CYCLE_CHECK_MS) : Module(std::move(name), cycleCheckTime) {}
+    I2CModule(std::string name, unsigned int cycleCheckTime = DEFAULT_CYCLE_CHECK_MS) : Module(std::move(name), cycleCheckTime) {}
 };
